+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 - Added `craft.oauth` Twig variable
 - Added `craft.oauth.getAppByHandle()` Twig helper
+- Added CLI for refreshing app tokens `oauthclient/apps/refresh-tokens <app handle>`
+
+### Changed
+- refreshToken service method no longer accepts an $app parameter
 
 ## 2.0.2 - 2019-10-02
 ### Fixed